Simplify ItemRequirement alt usage
This was getting unecessarily complex without really adding anything useful.

This removes the addition of QuestItem and ItemDefinitions, and instead makes use of client.getItemDefinition for item details.

Also, this keeps the functionality allowing the usage of item names vs ItemRequirement name, but requires it be explicitely enabled by setting a boolean vs implicitly.